Fixed ScoreManager AddNewScore

+++ Assets/Scripts/ScoreList/ScoreData.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.SceneManagement;
@@ -12,7 +13,8 @@ using UnityEngine.SceneManagement;
 public class ScoreData
 {
     const string fileName = "C:\\Users\\emi_p\\OneDrive\\Desktop\\ScoreData.json";
-    public Score scores = new Score();
+    public Score scores = JsonConvert.DeserializeObject<Score>(File.ReadAllText(fileName));
+    private List<Score> scoresList = new List<Score>();
 
     public void AddNewScore(string name, int score)
     {
@@ -21,22 +23,31 @@ public class ScoreData
             names = name,
             score = score
         };
-        WriteNewScore(newScore);
+        scoresList.Add(newScore);
+
+        WriteNewScore(scoresList);
     }
 
-    public void WriteNewScore(Score newScore)
-    {        
-        string dataJSON = JsonUtility.ToJson(newScore);
-        File.WriteAllText(fileName, dataJSON);
-        SceneManager.LoadScene("HighScore");
+    //revisar
+    public void WriteNewScore(List<Score> newScore)
+    {
+        if(File.Exists(fileName))
+        {            
+            string json = JsonUtility.ToJson(newScore);
+            Debug.Log(json);
+            Debug.Log(scores);
+            File.WriteAllText(fileName, json);
+        }
+        else
+        {
+            string dataJSON = JsonConvert.SerializeObject(newScore);
+            File.WriteAllText(fileName, dataJSON);
+            SceneManager.LoadScene("HighScore");
+        }
     }
 
     public void ReadScore()
     {
-       string json = File.ReadAllText(fileName);
-        for (int i = 0; i < json.Length; i++)
-        {
-
-        }
+        Dictionary<string, string> nameDict = new Dictionary<string, string>();
     }
 }
